+++ projects/unit_test_merge_sort/src/merge_sort.hpp
@@ -5,8 +5,8 @@
 #include <algorithm>
 
 namespace MergeNameSpace {
-	void merge(std::vector<int>& vec, int l, int mid, int r);
-	void merge_sort(std::vector<int>& vec, int l, int r);
+	void merge(std::vector<long long>& vec, int l, int mid, int r);
+	void merge_sort(std::vector<long long>& vec, int l, int r);
 }
 
 #endif
